+++ packages/cketh/candid/minter.certified.idl.js
@@ -40,6 +40,12 @@ export const idlFactory = ({ IDL }) => {
     'stopping' : IDL.Null,
     'running' : IDL.Null,
   });
+  const QueryStats = IDL.Record({
+    'response_payload_bytes_total' : IDL.Nat,
+    'num_instructions_total' : IDL.Nat,
+    'num_calls_total' : IDL.Nat,
+    'request_payload_bytes_total' : IDL.Nat,
+  });
   const DefiniteCanisterSettings = IDL.Record({
     'freezing_threshold' : IDL.Nat,
     'controllers' : IDL.Vec(IDL.Principal),
@@ -49,6 +55,7 @@ export const idlFactory = ({ IDL }) => {
   const CanisterStatusResponse = IDL.Record({
     'status' : CanisterStatusType,
     'memory_size' : IDL.Nat,
+    '_stats' : QueryStats,
     'cycles' : IDL.Nat,
     'settings' : DefiniteCanisterSettings,
     'idle_cycles_burned_per_day' : IDL.Nat,