+++ packages/cmc/candid/cmc.certified.idl.js
@@ -18,11 +18,16 @@ export const idlFactory = ({ IDL }) => {
     'Filter' : SubnetFilter,
     'Subnet' : IDL.Record({ 'subnet' : IDL.Principal }),
   });
+  const log_visibility = IDL.Variant({
+    'controllers' : IDL.Null,
+    'public' : IDL.Null,
+  });
   const CanisterSettings = IDL.Record({
     'controller' : IDL.Opt(IDL.Principal),
     'freezing_threshold' : IDL.Opt(IDL.Nat),
     'controllers' : IDL.Opt(IDL.Vec(IDL.Principal)),
     'reserved_cycles_limit' : IDL.Opt(IDL.Nat),
+    'log_visibility' : IDL.Opt(log_visibility),
     'memory_allocation' : IDL.Opt(IDL.Nat),
     'compute_allocation' : IDL.Opt(IDL.Nat),
   });