+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Upgrade candid files for ic-managmenet canister and support new field `reserved_cycles_limit`.
 - Add "API Boundary Node Management" topic support.
 - Add optional field `logo` to `Token` type.
+- Update `sns-js` candid files with new fields in sns canisters.
+- Add public method `toE8s` to `TokenAmountV2`.
 
 ## Breaking changes